Fix CTAD warning.
Without specifying the type explicitly, we get:
```
src/ducc0/infra/threading.cc:651:7: error: 'ScopedValueChanger' may not intend to support class template argument deduction [-Werror,-Wctad-maybe-unsupported]
```
The solutions are either to explicitly specify the tool (simply `bool` in this case), and/or add deduction guides.  The former was easier for now, but we can
certainly add deduction guides later if needed.